requirements.txt
pyobjc-core
